+++ design-system/radio/examples/02-form-example.tsx
@@ -18,6 +18,17 @@ const fruitItems: OptionsPropType = [
   { name: 'fruit', value: 'peach', label: 'Peach' },
   { name: 'fruit', value: 'pair', label: 'Pair' },
 ];
+const cityItems: OptionsPropType = [
+  { name: 'city', value: 'sydney', label: 'Sydney' },
+  { name: 'city', value: 'mountain-view', label: 'Mountain View' },
+  { name: 'city', value: 'new-york-city', label: 'New York City' },
+  { name: 'city', value: 'gallifrey', label: 'Gallifrey', isDisabled: true },
+];
+const weatherItems: OptionsPropType = [
+  { name: 'weather', value: 'sunny', label: 'Sunny', isDisabled: true },
+  { name: 'weather', value: 'cloudy', label: 'Cloudy', isDisabled: true },
+  { name: 'weather', value: 'windy', label: 'Windy' },
+];
 
 export default function FormExample() {
   return (
@@ -52,6 +63,32 @@ export default function FormExample() {
               >
                 {({ fieldProps }: { fieldProps: object }) => (
                   <RadioGroup {...fieldProps} options={fruitItems} />
+                )}
+              </Field>
+              <Field
+                label="radio group with individual field disabled"
+                name="city"
+                defaultValue="sydney"
+              >
+                {({ fieldProps }: { fieldProps: object }) => (
+                  <RadioGroup
+                    {...fieldProps}
+                    isDisabled={undefined}
+                    options={cityItems}
+                  />
+                )}
+              </Field>
+              <Field
+                label="radio group with individual field enabled"
+                name="weather"
+                defaultValue="windy"
+              >
+                {({ fieldProps }: { fieldProps: object }) => (
+                  <RadioGroup
+                    {...fieldProps}
+                    isDisabled={undefined}
+                    options={weatherItems}
+                  />
                 )}
               </Field>
               <FormFooter>